mymp/requirements.txt
alabaster==0.7.12
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.5.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-configurations==2.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-timezone-field==4.1.2
docutils==0.16
flake8==3.9.1
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.21
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
sorl-thumbnail==12.7.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
twilio==6.56.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
